// src/app/models/security/register.model.ts
export class RegisterModel {
    id: number = 0;
    username: string = '';
    name: string = '';
    email: string = '';
    user_type: string = 'User';
    warehouse_id: number = 0;
    warehouse: string = '';
    is_active: boolean = false;
    password: string = '';
    password_confirmation: string = '';
}